生成スピード調整処理落ち防止
